refactor: split announcement tasks
